requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.2
email-validator==2.2.0
Werkzeug==3.0.3
WTForms==3.1.2
Flask-Admin==1.6.1
python-dotenv==1.0.1
